Gave the beam break sensors their own subsystem
